+++ src/main/java/com/saggezza/blackjack/ICalculateBet.java
@@ -1,5 +1,5 @@
 package com.saggezza.blackjack;
 
 public interface ICalculateBet {
-     double calculate( double money,int bet,String gameResult);
+     double calculate( double money,double bet,String gameResult);
 }
